bug fix in code execution
TODO: need to fix pb in mapelite sometimes: "if fitness >= self.fitnesses[map_ix]:" at l546 crash
fix bug  in p3 env in the generation of program, when the execution of one puzzle generated doesn't work it relaunched a new generation of all others program. The execution is to get the output of g() however when g output an object it seems to crash. I added try and except to not get an error and set result_obj (output of g()) to "Error" when the execution failed.